Fix ResourceWarning in test_build_ext_config_handling
The test was failing on Cygwin with Python 3.9 due to unclosed file
descriptors in subprocess.Popen. This fix ensures that stdout and stderr
pipes are properly closed even when proc.communicate() is called.

The issue manifested as:
  pytest.PytestUnraisableExceptionWarning: Exception ignored in: <_io.FileIO [closed]>
  ResourceWarning: unclosed file <_io.BufferedReader name=16>

This fix wraps the communicate() call in a try/finally block to ensure
the pipes are always properly closed, preventing the ResourceWarning.